requirements.txt
PyQt6
python-mpv
requests
imdbpy
